Moved Event models into model package
