+++ app/src/main/java/com/cmput301f19t09/vibes/fragments/moodlistfragment/MoodListFragment.java
@@ -9,19 +9,17 @@ import android.widget.ListView;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
+import com.cmput301f19t09.vibes.EditFragment;
 import com.cmput301f19t09.vibes.MainActivity;
 import com.cmput301f19t09.vibes.R;
 import com.cmput301f19t09.vibes.models.MoodEvent;
 import com.cmput301f19t09.vibes.models.User;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class MoodListFragment extends Fragment
 {
     private ArrayAdapter<MoodEvent> adapter;
     private List<MoodEvent> data;
-    private final MainActivity parent = (MainActivity) getActivity();
+    private final MainActivity parent_activity = (MainActivity) getActivity();
 
     private String username;
 
@@ -45,30 +43,21 @@ public class MoodListFragment extends Fragment
 
         username = (String) arguments.getSerializable("username");
 
-        ListView view = parent.findViewById(R.id.mood_listview);
+        ListView view = parent_activity.findViewById(R.id.ml_listview);
 
         view.setOnItemClickListener(new AdapterView.OnItemClickListener()
         {
             @Override
             public void onItemClick(AdapterView<?> self, View view, int position, long id)
             {
-                Bundle arguments = new Bundle();
 
                 MoodEvent event = (MoodEvent) self.getItemAtPosition(position);
 
-//                arguments.putSerializable("edit", true);
-//                arguments.putSerializable("mood", event);
-
-//                ((MainActivity) parent).replaceFragment(AddFragment.class, arguments);
+                parent_activity.setMainFragment(EditFragment.newInstance(event));
             }
         });
 
         data = new ArrayList<MoodEvent>();
         adapter = new MoodListAdapter(getContext(), data);
-    }
-
-    private void loadData()
-    {
-
     }
 }
