+++ app/src/main/res/layout/mood_list.xml
@@ -16,7 +16,7 @@
     </FrameLayout>
 
     <ListView
-        android:id="@+id/mood_listview"
+        android:id="@+id/ml_listview"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
